Finished exercise for Simple Todo List app.
This is from ZTM Complete Webdev, Lecture 142. I must say I'm proud of
what I did; I wasn't expecting the exercise to be this interesting!
Then again, practice makes perfect, and experience begins to compound
on itself. But I digress. :)

I took a few liberties with the HTML layout, mainly encapsulating the
li element with its delete-button in a separate container div with
class "entry". This way, I could style the delete-button's position
more appropriately, using in fact a technique we learned earlier in
the course (using an auto-margin with a flexbox item).